Removed unused client-side DAOs. These have been replaced by client-side handler implementations.
On the server, there have been to date an (ideal) seperation of several layers:
- Persistence: (hibernate) Handled all database interactions, mapping between database and domain objects
- Data access layer: meant to depend soley on persistence layer, expressed (ideally) in JPQL
- Bean mapping: (dozer) provided mapping between domain objects and data transfer objects
- CommandHandlers: ideally meant to depend only on Data access layer and Bean mapping layer

The strategy had been to replace the persistence layer with a compile-time implementation of the JPA (rebar-persistence) and parts of the Data access layer as needed (e.g. for the SiteTableDAO which used the hibernate criterion API)

As it turned out, this introduced more problems than it solved. Lazy loading of child objects, for example, is straightforward to implement, but not very useful until you optimize for subselects, (harder!) otherwise you end up with the n+1 query problem. And then we still have to map our domain object to DTOs on the client-side...

Anyway, writing CommandHandlers against JDBC + manual DTO mapping is tedious, but it doesn't take too long to write, are easily testable, and are quite efficient client side. (Faster, in fact, than some of the server side handlers that rely on hibernate for relational gymnastics)

We need to get offline working for the folks in the DRC who have been waiting so long, so we're going to stomach the duplicaton of client + server handlers and just get it working, and review the design at a later date in light of concrete working classes. 




git-svn-id: http://sigma-h.googlecode.com/svn/trunk@424 46f870f5-83fa-7f44-208d-45d3d0f2e120