I think this is the problem..... The idea is to break up Locations to be sent to the client for sync in batches of 25.. so we order by the dateEdited and use dateEdited as a cursor bookmark.
However, it can happen for one reason or another that a large number of entities share the same dateEdited value. So impose a total ordering by incorporating the primary key and use (dateEdited, id) as the cursor bookmark. But some how it gets all screwed up... 

I think this query as it was was wrong, but not sure. Frustratingly, the unit tests pass, it only breaks down with real world data.

Let's see if this works.
Any one else see an error?


git-svn-id: http://sigma-h.googlecode.com/svn/trunk@659 46f870f5-83fa-7f44-208d-45d3d0f2e120